Maak EetplanService
Trek logica uit EetplanRepository naar een specifieke klasse die eetplannen maakt
